feat: show crafted-by footer in header when sidebar is collapsed
Lift the sidebar collapsed state up to DashboardShell so the Sidebar and
Header share it. When the sidebar is collapsed (desktop), the
"Crafted with… by Sanghel González" footer moves from the sidebar into
the header. On mobile it stays hidden.

Closes #495
